Disaster/ontology_service/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
rdflib==7.0.0
pandas==2.1.4
numpy==1.26.2
Pillow==10.1.0
openpyxl==3.1.2
